requirements.txt
asgiref==3.5.2
Django==4.0.5
django-crispy-forms==1.14.0
django-debug-toolbar==3.4.0
django-tinymce==3.4.0
djangorestframework==3.13.1
Pillow==9.1.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
